// apps/step0110/step0110-e2e/src/enums/TestIdEnumeration.ts
export enum TestIds {
  SHOPPING_CART_BTN = "open-shopping-cart-button",
  ADD_TO_CART_BTN = "add-to-cart-button",
  ITEM_CART_ENTRY = "item-cart-entry",
  BUY_NOW_BTN = "buy-now-btn",
  PERSONAL_INFORMATION_NEXT_BTN = "stepper-personal-information-next-button",
  VORNAME_INPUT = "order-vorname-input",
  NAME_INPUT = "order-name-input",
  STRASSE_INPUT = "order-strasse-input",
  PLZ_INPUT = "order-plz-input",
  ORT_INPUT = "order-ort-input",
  ORDER_NOW_BTN = "order-now-button",
  LOGO_IMG = "logo",
  NOT_FOUND_BTN = "not-found-back-to-home-button",
  NOT_FOUND_HEADER = "not-found-header",
  SHOP_ITEM = "shop-item",
  PERSON_ID_INPUT = "person-id-input",
  PERON_SEARCH_BTN = "person-search-btn",
  LOGIN_BTN = "login-btn",
  PASSWORD_INPUT = "passwort-Input",
  USERNAME_INPUT = "username-Input",

}